ems-checking/requirements.txt
numpy
fastapi==0.89.1
uvicorn==0.20.0
tensorflow
scikit-learn
opencv-contrib-python-headless==4.5.3.56
h5py
matplotlib
Pillow
requests
imageio
openvino
pandas
pyjwt[crypto]